更新JQLibrary

+++ lib/JQLibrary/src/JQQRCodeReader/zxing/zxing/oned/CodaBarReader.h
@@ -26,8 +26,8 @@ namespace oned {
 
 class CodaBarReader : public OneDReader {
 private:
-  static const int MAX_ACCEPTABLE;
-  static const int PADDING;
+  static const float MAX_ACCEPTABLE;
+  static const float PADDING;
 
   // Keep some instance variables to avoid reallocations
   std::string decodeRowResult;
