+++ lib/JQLibrary/src/JQQRCodeReader/zxing/zxing/oned/Code128Reader.h
@@ -26,8 +26,8 @@ namespace oned {
 
 class Code128Reader : public OneDReader {
 private:
-  static const int MAX_AVG_VARIANCE;
-  static const int MAX_INDIVIDUAL_VARIANCE;
+  static const float MAX_AVG_VARIANCE;
+  static const float MAX_INDIVIDUAL_VARIANCE;
 
   static std::vector<int> findStartPattern(Ref<BitArray> row);
   static int decodeCode(Ref<BitArray> row,
